requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
confluent-kafka==1.8.2
dateparser==1.1.1
frozenlist==1.3.0
gevent==21.12.0
greenlet==1.1.2
idna==3.3
jsonschema==4.7.2
mongoengine==0.24.1
multidict==6.0.2
py4j==0.10.9.3
pycodestyle==2.8.0
pymongo==4.1.1
pyrsistent==0.18.1
pyspark==3.2.1
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.27.1
six==1.16.0
toml==0.10.2
tzdata==2022.1
tzlocal==4.2
ujson==5.3.0
urllib3==1.26.9
websocket-client==1.3.2
websockets==10.3
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
